Permissions for creating and editing lists.
Change the firebase rules for list creation and editing, to be less
specific, but still as secure.

In the app, when a list is created, add the user to the list of editors
of that list, so that the user has the firebase permissions to save the
list's initial action list.