Fixed oversubscriptions and added a lot of debug output
In a flurry of confusion, add logging to all page loads everywhere.

Eventually conclude that the important thing is that anywhere we
set up global state, we must tear it down with onDestroy to keep
things clean.

This means things like auth callbacks, onSnapshot callbacks, and
even potentially things in the redux store.

For the store, handle it by trapping sign out and cleaning out all
state. This might not really be good enough if someone causes action
replay, but since what causes that is signing out and signing in
again at the moment it is working fine.