Set page title to name of list.
The page title is set by the ItemDisplay when it is
clicked on, and so it knows the correct name.  That
name is set via a new ui reducer that is meant only
for local settings not to be sent to the firebase server.
Then the title is read and displayed by the layout.