Close dialogs with ESC.
If a dialog uses state from $store to know if it should be visible or
not, it can't be closed by pressing ESC or clicing behind the dialog.

Add a new class SgDialog that uses a closeHandler that detects this
case and calls a supplied cancelDialog function.  The function is
resposible for changing the state with an action.