First cut of dragging lists to rearrange.
We gave up on making the draggable list generic and just
cloned the code and pruned it slightly to make it "work".

Two gotchas for the future:

1- the anchor shouldn't use its parent element in the new version,
but needs to in the old version because the text field above it also
scrolls and uses up space in the scrolled div. When we fix that (as
we probably don't really want that field to scroll) it semes likely
that this will be wrong and need to match the new code and use
the anchor's position directly.

2- when you drag off the bottom of the list of lists, the mouse
events get lost and you can easily wind up in a bad state.

Other than that, this change looks like it works!