Don't redirect if we don't know our sign-in state.
When the sign in callback hasn't happened yet, if we don't redirect
then when it does happen if the user is signed in the expected page
will refresh and it all looks good.

Clearly good parts of this change involve setting signedIn to
undefined at system startup, so that we can understand the three
states: we don't know, we are definitely signed in, we are
definitely signed out. Some thought/experimentation has to be done
with offline to see what state we are in when offline and ensure
that the rest of the code does the right thing in that case.

For now, this seems to work even if we don't fully understand the
implications.