Check if item.description changed.
Don't issue an event for a no-op in the case where an item's
description is edited, but not changed.  This can happen when
clicking around in a list and will cause an ItemDisplay to go
into edit mode (ie: it shows an edit cursor).  Clicking elsewhere
without making an edit fires the blur we use to detect edits to
the description.

Now, check if the new description is different from the original
description, and only issue an event if it is.