Fix list not updating behind dragged item.
Oh wow!  Not sure what the $: is pulling in to update
and how that's related to the actual list, but limiting
the scope of $: to one variable is the key.

The problem was that the list items displayed behind
the dragged item don't change position as the dragged
item moves over.