A FilterMenu item is active if url matches its name.
The FilterMenu items are the list names like 'All', 'Starred', etc.

Before this change, there was no indication which FilterMenu was
currently active because the background of the one in use wasn't
changed.

Set a menu item 'active' so it can draw with a different background
colour to show that it is the item that is selected.  This is done
by looking at the url, and seeing if it ends with '/all', or
'/starred', etc.