Make mobile 'enter' key submit.
The mobile 'enter' key's default for textfields seems
to be "next" which will move the focus to the next
textfield on the page.  We want it to be treated as
a keydown so we can detect it.  We do this in two
places: to create a new list, and to create a new item.

A form's input element can have an enterkeyhint for
exactly this purpose, which tells a mobile keyboard
what it might display on its enter key (eg: 'search').
We'll set the hint to 'enter' which sends the
keydown event we want.