Fix ghost in ItemList: correct y pos for long list
When there are enough list items for a scroll bar, the ghost is not
drawn in the correct position, and is off by the amount that the window
is scrolled.

We had the same issue with the ListMenu, but we didn't fix it here.
Indeed, this fix is duplicated from that code.