Go back to our original colour scheme.
After a bunch of futzing we couldn't easily get dark mode buttons
to render well and it seemed easiest to persist with just hardcoding
light mode for everyone.

The original colour scheme that uses the same background colour as
the component colour was our preferred option so we undid that change,
and the only visible change here from the previous light mode is that
the text input fields are tinted because we removed all hardcoded
background colours.

The only hardcoded colours that remain are those that define the
red and dark text for due dates.