Bugfix: Completed page filters properly.
In an attempt to show the recurring items that had been completed at
any time in the past, the filter only looked at the compltedTimestamp.
This caught too many things, including normal (non-recurring) items
that were completed then uncompleted.
The filter now matches items that are completed, and recurring items
that have been completed at some time in the past.