Offline support.
Turn on offline support by configuring firestore to work locally.

All actions are given to us by firestore via two paths: server actions,
and local actions.  We now execute all of these actions all the time.
When we are on-line we get two copies of the same action, which now
means that our actions must be idempotent.