Implement a rebasing reducer for offline.
When the user is offline, our strategy is to play offline actions
onto the end of the server side state, waiting for things to go
online again later. This is done by implementing a new root reducer
that keeps a separate redux store of the server side state, and
a list of offline actions to be rebased on top of it. This new
root reducer removes items from the local list as they appear from
the server, so that each action is executed exactly once.

This solution removes the need for idempotent actions which turned
out to be trickier to accomplish than we first thought, and instead
provides the guarantee that each action will be executed once against
the state.

Some oddnesses: we didn't manage to make the types clean without
having the internal serverSideStore needlessly be a Svelte Store,
even though nothing ever subscribes to it; and the maxAge parameter
seems to be passed to the wrong store but affects redux dev tools
in the right way so we just left it there instead of figuring out how
to set that option on the "correct" create store call.

Also separate from everything else in this change we observed that
some actions like completing items get done twice and this is
technically harmless but was unexpected.