A finally working android app.
After a large number of fumbles and fixes, we eventually got auth
working in the Android capacitor app version.

The main thing seemed to be a missing dependency to get
GoogleSignIn installed by adding the exact right magic lines to
android/app/build.gradle. Additionally, the last magic we seemed
to need was one or both of adding rgcfaIncludeGoogle to the
anroid/variables.gradle file and putting the SHA-1 key of the app
into the firebase console. Up until then, the app was complaining
that it couldn't find a capacitor plugin to field the request.