Create a new ios capacitor app.
This is the result of runnign the commands:

npm i @capacitor/ios
npx cap add ios
npx cap sync

and then firing up xcode and building the project.

The resulting project does not work because we haven't yet
done the manual steps required to get the firebase-auth capacitor
plugin to work properly. The symptom at this stage is a crash that
starts with the somewhat helpful log message:

2023-03-08 12:21:54.800248-0500 App[483:21570784] *** Terminating app due to uncaught exception 'com.firebase.core', reason: '`FirebaseApp.configure()` could not find a valid GoogleService-Info.plist in your project. Please download one from https://console.firebase.google.com/.'

which appears in a panel in XCode when you launch the app in the
emulator.