Add GoogleService-Info.plist to the xcode project.
First, go to the firebase console to add an iOS app, which
as part of its setup instructions here:

https://console.firebase.google.com/project/todo-firebase-1a740/settings/general/ios:com.stockgamblers.todo

tells you to download and add the plist file to the xcode project.

This resolves the error message from the previous commit that more
or less tells you to do this step.