Add FirebaseApp.configure() to iOS app.
When running the app, firebase was complaining that it wasn't
initialized with a message like this in the XCode console:

2023-03-08 12:53:42.421855-0500 App[3994:21605876] 10.2.0 - [FirebaseCore][I-COR000003] The default Firebase app has not yet been configured. Add `FirebaseApp.configure()` to your application initialization. This can be done in in the App Delegate's application(_:didFinishLaunchingWithOptions:)` (or the `@main` struct's initializer in SwiftUI). Read more: https://goo.gl/ctyzm8.

Add the required call in this change.

For each rebuild, it should suffice to do:

    npx cap update
    npx cap sync
    npx cap open ios

and rebuild inside Xcode's GUI. Note: if you are having a problem
where you click the sign-in button and it doesn't connect with the
native sign in, it seems to b because pod install wasn't executed
and that is done by the 'npx cap update' step above. This often
won't be needed so you'll forget.