Update app.html and (app)/+layout.svelte for iOS.
Use env(safe-area-inset-top) and friends in the CSS to minimally
alter the UI to avoid crashing into the notch.

This deliberately works only for the case where the user is holding
their phone in portrait, or in landscape rotated 90 degrees
counter-clockwise, but not in the other two orientations.

The top app bar adds padding to fill the notch area/curves with the
brown top app colour, and the drawer adds padding on the left, top
and bottom to avoid colliding with anything.