Awful hack to make orientation work on iOS.
For reasons unknown, binding the inner width doesn't work on
capacitor iOS (though it does in safari itself). Set up an
event handler for onorientationchange and write the width
variable to force it to update.

Still that proved not enough and the markup itself needed to
be touched to force the re-render, so add a dummy w<width>
class to the top level div which does nothing other than
force the page to lay out again and display the drawer.

Awful, low-level of true understanding hack that "works".