Bugfix: EditDetailsDialog can show old info.
The bug: While editing the description of a todo item, press the Details
button to view the Details Dialog.  The dialog will show the description
from before you started editing.

The bug was caused because the code that updates the values for the
dialog (description, due date, etc) only checked that the itemID hadn't
changed, and assumed that if it's the same id, it's the same data.

The fix is to not rely on the previous itemID, and to do a deep equals
on the object to detect if any properties have changed.