Bugfix: Closing share dialog revokes all shares.
Previously, when the dialog was opened, the list of shared users
was always empty.

Initialize the list of shared users before showing the sharing dialog.
When the dialog is closed, the correctly initialized list can be
compared correctly to the list the user edited.

Related to this change:  Fix a problem in shareAccepted that now checks
that the last request is of the type 'accpet_pending_share', which is not
to be confused with 'revoke_share'.