WIP: A test to profile slow reducers.
We can now conclusively blame immer for the slowness of the 'items'
reducer.  This test ran in about 150ms with immer, then in less than
20ms without.

The items reducer in this change has been converted to sidestep immer,
where the system uses immer, but the reducer gets the original state
from immer, modifies it, then freezes the state tree before returning
it.

A possible next step is to write a drop-in replacement for
createReducer() so we can switch our reducers between using immer and
not without changing the reducers themselves.