More responsive Search field for short words.
The search field searches as soon as a character is typed, and that is
responsive when there aren't many matches.  When the number of matches
is larger, the drawing time is longer and typing more characters is
interrupted (especially on mobile).

This change adds a short delay between typing and when the search
starts, which is interrupted if a character is typed soon enough.  This
is applied to search terms of 3 characters or less.