Load & process actions before UI paints.
Use the load() function of src/routes/(app)/+layout.ts to load and
process all actions from all lists.  This load() function runs after a
user is authenticated, but before the UI is drawn.

From here we can load all actions from firestore and process them into
the redux state, as well as communicate to the UI.  We can control how
much of the UI is drawn, from the full UI (which was the case before
this change), to just the word "waiting" (which is how this
yet-to-be-finished change has left it.

Once the actions have been replayed, the UI draws the complete state and
is ready to go.

The average time to load and process our sample data on local firestore
is now about 6500ms.  The UI is completely drawn shortly after that, say
at 6700ms.