Bugfix: item is undefined after switching lists.
Steps to reproduce:

  1. Choose any list item and open its details dialog.
  2. Switch to another list by clicking on its name.

Result:

  _layout.svelte-9b5e5290.js:484 Uncaught (in promise) TypeError: Cannot
  read properties of undefined (reading 'description')

The code that keeps the information in the dialog up-to-date watches for
changes to $store.ui.itemId and $store.ui.listId, then looks up that
itemId in $store.ui.listId.  But these are not necessarily in sync, so
when the listId changes, the code can't find the itemId from the
previous list in the new list.

So don't try to update the dialog if the item can't be found in the
current list.

Also make the $: block clear that we want to updated only on changes to
  $store.ui.listId and
  $store.ui.itemId
by moving the update code into a function that is called like this:

  $: updateDetailsDialog($store.ui.listId, $store.ui.itemId);