First cut of caching the state in indexed db.
It became clear that replaying actions from firebase was never
going to be fast enough on mobile as just querying the local
cache for actions was taking something ridiculous like 5s on
the browser on android and even longer inside the capacitor app

So, cache the realized state along with the timestamp of the last
action definitely executed against it in a new indexed db instance
named "state" with a key of the user's email and a value of the
entire state object. E.g. if two users alice and bob are using the
device you'd have alice@gmail.com -> "entire state for alice" and
likewise for bob.

Then, upon first load, try to load the state from indexed db only
replay the actions whose timestamps are later than the latest timestamp
used to construct the saved state. In order to avoid any edge cases
with duplicate timestamps, the code which writes the state is careful
to only save state where the timestamp is strictly less then the timestamp
of the currently about-to-be-processed action; while this means the
cache is always possibly somewhat behind, it also means that the
timestamps cannot overlap.

Once the state is loaded, if there is a non-zero timestamp from the
cache go ahead and display what we've got while the database actions
are loaded and (if new enough) executed.

This works more-or-less OK on desktop. On mobile, using a where clause
to filter the incoming actions created noticable delays, which we think
affects desktop too by eliminating client-side echo of the actions as
the user works. Probably filtering in javascript like we've done in
this change is the right answer but more testing is needed to see how
that affects load time performance.

Performance is still not really acceptable on mobile, and on desktop
the UI still renders a bit slow (the by date view takes > 1s to render
because of the complexity of ItemDisplay svelte components). The next
priority ought to be to see if we can get android to load in non-glacial
time.

Things to check:
- we shouldn't be waiting on the state saving before processing
the in flight action;
- and the fact that we didn't wait shouldn't mean we write a later
version of the state by accident.