Watch pending shared lists to get list name.
When the list of pending shares is changed, watch the list to get its
actions, including its name.  This changes the list name from
'undefined' to the correct name.

We don't remember why list items don't appear when a pending share is
clicked on.  That's the behaviour we want anyway (we just don't
remember why).