Add button to complete a recurring item forever.
Add a button at the top level that completes a recurring item.  This
does replaces the previous job of going into the details dialog, marking the
item as not recurring, closing the dialog, then checking the checkmark.

As a bonus, undoing a recurring item is now possible, while retaining
the recurring status and values.