Make combobox items shorter, to display all items on phone.
The Edit Task dialog is as tall as it can reasonably be for some phones,
allowing for a keyboard to be desplayed beneath it.  The full list of
options in the Repat combobox is somehow too tall, and often the
first item isn't displayed.  This is related to our unhappiness with
the combobox, as there is room in the space of the dialog to display
all of the options.

Make each item in the Repeat combobox smaller so when expaneded,
the list of all options is shorter.