Timestamps for each list & cache.
Have a timestamp for each list which tells us the timestamp of
its last action.  Use this to filter all actions coming from the server
to reduce replay time.

Rework caching of the app's state to cache the one-action-old state
when a new server action arrives, but wait one second before writing
the cache.