Avoid updating per-list timestamp from global action.
Previously, reorder list actions containing a list id would
be seen by the reducer and cause the timestamp per-list to
update. However these global actions should only affect the
global timestamp, not the filter on the per-list actions, so
the reducer now ignores the global actions which are already
flagged as "not a normal action".