Organize startup so that initial actions happen once.
We turned out to be pretty confused about startup order/startup
functions being called once or multiple times.

In a few places we have a pattern of setting a variable from
undefined to a defined value, and "protecting" against executing
the code twice by checking the variable against undefined. With
the advent of using async a bunch at startup, we created "race"
conditions where the code could be re-entered before the variable
was initialized by code after an async await; fix these by setting
the locking variable to 'null' immediately to avoid re-entrancy
issues.

Also, on the phone, we found that it was possible to get subscribed
to auth messages multiple times, because we had a subscription for
load time that was never unsubscribed. Fix it by hooking it up to the
existing cleanup code.

All this compelled us to as our next task go back and understand
every nuance of startup once again, and perhaps we'll be able to
clarify this apparently quite confusing code.