Upgrade to capacitor 5.0.0.
Upgrade many packages to satisfy the compiler so we can build after
installing capacitor 5.0.0.

To upgrade capacitor, we did:
  npm remove @capacitor-firebase/authentication
  npm install vite@latest
  npm update
  npm i -D @capacitor/cli@latest
  npx cap migrate
  npm install @capacitor-firebase/authentication

Then we ran into several dependency issues with svelte, typescript,
smui, and their various dependencies. Two specific surprises were that
we had to install @typescript/svelte manually to satisfy smui, and having
done that we found that the configs included then required typescript 5.
Along the way we removed and re-installed pretty much all of svelte and
@smui to try to ensure we were on the latest packages, but how much of
that was needed is unclear.

Also we weren't sure if the @smui packages were installed consistently
with -D to be on the development dependencies list, but we tried to be
good about doing that correctly for all these installs.

Looks like we might still have packages on the prod side that maybe
should be dev dependencies, namely:

    "@capacitor/android": "^5.0.0",
    "@capacitor/core": "^5.0.0",
    "@capacitor/ios": "^5.0.0",
    "@sveltejs/adapter-static": "^2.0.3",
    "svelte": "^4.1.2",
    "svelte-check": "^3.4.6",
    "svelte-preprocess": "^5.0.4",
    "typescript": "^5.1.6"

With some of these we're not sure and others (like typescript)
almost surely should be dev only.