Serverside: Send notifications to all editors.
When the server notices a todo item has changed, it
now looks up the message tokens for all editors of
the list, and sends a notification to each.

Next steps:

- update the server side to maintain a list of FCM tokens,
  invalidating the if push requests fail (which maybe is the
  perfect strategy or maybe not). We can watch the users collection
  for updates to accumulate new incoming tokens without modifying
  the client side, and clear out bad tokens when notifications
  are sent but fail.

- find out what the special keys are per platform for images, text
  etc to make pretty notifications for the cases we care about

- build an iOS version and see if it works

On a non-notifications note, test new accounts and make it not
stay stuck on "Initializing" status. Just a fix to the fix that
we already put on main that updates the loading status in that
code path.