Fix: Daily repeating items don't show repeat char.
This bug was introduced when we shortened the date display to have no
year if the date was in the current year.  The regex used for inserting
the repeat character for the daily case depended on the year being
present.

We known if the year is going to be there or not, so we can insert the
repeat character the old way when it is, or simply append the repeat
character to the end of the date if there is no year present.