Collect notification tokens from multiple devices.
Collect all notification tokens that come from a user's many devices.
When a login happens, a new possibly new notification token is written
to the server database.  The server-side notices this change and adds
the token to a list of tokens known for this user.

When sending a notification to a user, read all of these tokens to send
to all devices.