Enable caching after load completes.
When we know that all the data is loaded, enable caching of the local
state.  This avoids the problem that a cache could be written while
lists are loading, which would have the cache contain an incorrect cache
timestamp (because each list is replayed from the beginning of time,
one after the other).

Also be less conservative on loading list actions. Previously we'd
allow for duplicate actions on the latest timestamp of every list
because the timestamps are on per-second granularity; now only allow
for that if there is a list whose timestamp equals the overall cache
timestamp (which happens almost never). The net effect of these two
fixes is that normally the UI can load and look complete immediately.