Add a new "Today" list that has starred and due items.
Alex wanted to integrate some new daily habits and rather than
use the stars for that exclusively, there is now a new "Today"
list that shows all starred items first in their proper order
followed by items that are due today (or overdue). The newly
pushed-on-you overdue items can be starred to promote them to
the top of the list or ignored.

While we were at it, ItemList used to have a sort of hardcoded
notion of which lists should display list names, so we made
that a parameter for easier re-use from the new screen, and
made the old ones pass it.

In Alex's VSCode there are new problems shown in by the
typescript checking, and we fixed one which was an out-and-out
typo of the word "length". This looked like it had a potential
performance impact but upon investigation turned out to be
harmlessly wasteful.

More typescript fixes to come in a follow-up CL.